Change timer from milliseconds to nanoseconds to fix bug where sometimes divides by zero. Use float instead of int where possible

+++ headers/pid_controller.h
@@ -1,10 +1,10 @@
 #ifndef PID_CONTROLLER_H_
 #define PID_CONTROLLER_H_
 
-int pid_compute(int input);
+float pid_compute(float input);
 void set_pid_parameters(float P, float I, float D);
-void set_pid_set_point(int value);
-void set_pid_output_limits(int min, int max);
+void set_pid_set_point(float value);
+void set_pid_output_limits(float min, float max);
 
 #endif /* PID_CONTROLLER_H_ */
